requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
Pillow==10.4.0
transformers==4.45.2
torch==2.4.1
python-multipart==0.0.12
